Adding support for dirty checking
On every single edit page, we should do dirty checking to ensure that
changes are not being lost. I added a nice directive that really
simplifies things. Just add the “unsaved-warning-form” attribute to
your form, and it handles the rest. You can see it in action on the
blog edit page.