Adding initial support for PouchDB
- right now the datastore is doing a check against local storage for
all objects and returning them as a promise if they exist. Huge
performance gains. The use of $timeout when returning the local storage
object is hacky, but it’s the best way I could think of to return a
synchronous call asynchronously. The only reason I’m doing this is
because the controllers expect an async call when getting an object
from the datastore

- all local storage support will move to pouchdb soon. It’s fully
supported right now, it just needs to be implemented in the services