Lots and lots of dataStore refactoring
Much efficiency

- The dataStore always checks localStorage when possible now. Each
category is saved locally, as is each item, but items first check
localStorage to see if their parent has been saved, otherwise it
filters through its parent by id and returns itself as an object

- data freshness is also implemented. Right now it’s only on the parent
categories, but that might be all we need it for. Each time a category
is saved locally, it’s associated with a timestamp. When the data is
fetched, it does a comparison against that timestamp. If the data is
more than an hour old, it fetches new data from the API