Adding pull to refresh
List pages now have pull to refresh functionality. When a user
refreshes a list, it forces a call to the API rather than getting data
from localstorage

IMPORTANT: This means you have to pass a boolean as a second param to
the getItemsAsync function. True means a call to the API is made. False
(which is the default) means the call goes through the process of
looking in localstorage first